requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.0
httpx==0.27.0
google-cloud-speech==2.26.0
aiofiles==23.2.1
aiohttp==3.9.5
pytz==2024.1
jinja2==3.1.3
python-multipart==0.0.9